+++ src/main/java/com/projectkorra/core/util/PluginUtil.java
@@ -6,12 +6,24 @@ import java.util.Date;
 import java.util.List;
 
 import org.bukkit.Material;
+import org.bukkit.block.data.BlockData;
+import org.bukkit.block.data.Levelled;
 
 public class PluginUtil {
 	
-	private PluginUtil() {}
+	private static List<String> weapons = Arrays.asList("axe", "sword", "pickaxe", "shovel", "hoe")
+
+	public static BlockData createWaterData(int level) {
+		Levelled data = (Levelled) Material.WATER.createBlockData();
+		data.setLevel(level);
+		return data;
+	}
 	
-	private static List<String> weapons = Arrays.asList("axe", "sword", "pickaxe", "shovel", "hoe");
+	public static BlockData createLavaData(int level) {
+		Levelled data = (Levelled) Material.LAVA.createBlockData();
+		data.setLevel(level);
+		return data;
+	}
 	
 	public static String getCurrentDate() {
 		return new SimpleDateFormat("yyyy/MM/dd HH:mm:ss").format(new Date());
